app_streamlit/requirements.txt

streamlit==1.40.1
python-docx==1.1.2
uuid==1.30
azure-ai-language-conversations==1.1.0
azure-keyvault==4.2.0
azure-identity==1.19.0
python-dotenv==1.0.1
